+++ zio-http/shared/src/main/scala/zio/http/codec/HttpContentCodec.scala
@@ -6,12 +6,14 @@ import zio._
 
 import zio.stream.ZPipeline
 
-import zio.schema.Schema
 import zio.schema.codec._
+import zio.schema.{DeriveSchema, Schema}
 
 import zio.http.Header.Accept.MediaTypeWithQFactor
 import zio.http._
+import zio.http.codec.internal.TextBinaryCodec
 import zio.http.internal.HeaderOps
+import zio.http.template._
 
 final case class HttpContentCodec[A](
   choices: ListMap[MediaType, BinaryCodecWithSchema[A]],
@@ -166,6 +168,82 @@ final case class HttpContentCodec[A](
 }
 
 object HttpContentCodec {
+
+  private final case class DefaultCodecError(name: String, message: String)
+
+  private object DefaultCodecError {
+    implicit val schema: Schema[DefaultCodecError] = DeriveSchema.gen[DefaultCodecError]
+  }
+
+  private val NameExtractor    = """.*<p id="name">([^<]+)</p>.*""".r
+  private val MessageExtractor = """.*<p id="message">([^<]+)</p>.*""".r
+
+  private val domBasedSchema: Schema[HttpCodecError] =
+    Schema[Dom].transformOrFail[HttpCodecError](
+      dom => {
+        val encoded = dom.encode
+
+        val name = encoded match {
+          case NameExtractor(name) => Some(name)
+          case _                   => None
+        }
+
+        val message = encoded match {
+          case MessageExtractor(message) => Some(message)
+          case _                         => None
+        }
+
+        (name, message) match {
+          case (Some(name), Some(message)) => Right(HttpCodecError.CustomError(name, message))
+          case _                           => Left("Could not extract name and message from the DOM")
+        }
+      },
+      {
+        case HttpCodecError.CustomError(name, message) =>
+          Right(
+            html(
+              body(
+                h1("Codec Error"),
+                p("There was an error en-/decoding the request/response"),
+                p(name, idAttr    := "name"),
+                p(message, idAttr := "message"),
+              ),
+            ),
+          )
+        case e: HttpCodecError                         =>
+          Right(
+            html(
+              body(
+                h1("Codec Error"),
+                p("There was an error en-/decoding the request/response"),
+                p(e.productPrefix, idAttr := "name"),
+                p(e.getMessage(), idAttr  := "message"),
+              ),
+            ),
+          )
+      },
+    )
+
+  private val defaultCodecErrorSchema: Schema[HttpCodecError] =
+    Schema[DefaultCodecError].transformOrFail[HttpCodecError](
+      codecError => Right(HttpCodecError.CustomError(codecError.name, codecError.message)),
+      {
+        case HttpCodecError.CustomError(name, message) => Right(DefaultCodecError(name, message))
+        case e: HttpCodecError                         => Right(DefaultCodecError(e.productPrefix, e.getMessage()))
+      },
+    )
+
+  private val defaultHttpContentCodec: HttpContentCodec[HttpCodecError] =
+    HttpContentCodec.from(
+      MediaType.text.`html`      -> BinaryCodecWithSchema(TextBinaryCodec.fromSchema(domBasedSchema), domBasedSchema),
+      MediaType.application.json -> BinaryCodecWithSchema(
+        JsonCodec.schemaBasedBinaryCodec(defaultCodecErrorSchema),
+        defaultCodecErrorSchema,
+      ),
+    )
+
+  val responseErrorCodec: HttpCodec[HttpCodecType.ResponseType, HttpCodecError] =
+    ContentCodec.content(defaultHttpContentCodec) ++ StatusCodec.BadRequest
 
   def from[A](
     codec: (MediaType, BinaryCodecWithSchema[A]),
